+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 2022-01-04
 ### Added
 - add `/usr/bin/capture-export` and a keybind to launch it
+- add `nm-connection-editor`
 
 ### Changed
 - remove a keybind to open `/installer` in the post-installation environment